+++ rust/bridge/shared/src/net/registration.rs
@@ -12,10 +12,10 @@ use libsignal_bridge_types::net::registration::{
 use libsignal_bridge_types::net::TokioAsyncContext;
 use libsignal_bridge_types::*;
 use libsignal_net::registration::{
-    CreateSessionError, ForServiceIds, RegisterAccountError, RegisterAccountResponse,
-    RegisterResponseBadge, RegistrationSession, RequestError, RequestVerificationCodeError,
-    RequestedInformation, ResumeSessionError, SessionId, SignedPreKeyBody, SubmitVerificationError,
-    UpdateSessionError, VerificationTransport,
+    CreateSessionError, ForServiceIds, NewMessageNotification, RegisterAccountError,
+    RegisterAccountResponse, RegisterResponseBadge, RegistrationSession, RequestError,
+    RequestVerificationCodeError, RequestedInformation, ResumeSessionError, SessionId,
+    SignedPreKeyBody, SubmitVerificationError, UpdateSessionError, VerificationTransport,
 };
 use libsignal_protocol::*;
 use uuid::Uuid;
@@ -24,19 +24,9 @@ use crate::support::*;
 
 bridge_handle_fns!(RegistrationService, clone = false, ffi = false);
 bridge_handle_fns!(RegistrationSession, clone = false, ffi = false);
-bridge_handle_fns!(
-    RegisterAccountRequest,
-    clone = false,
-    ffi = false,
-    jni = false
-);
+bridge_handle_fns!(RegisterAccountRequest, clone = false, ffi = false);
 bridge_handle_fns!(RegisterAccountResponse, clone = false, ffi = false);
-bridge_handle_fns!(
-    RegistrationAccountAttributes,
-    clone = false,
-    ffi = false,
-    jni = false
-);
+bridge_handle_fns!(RegistrationAccountAttributes, clone = false, ffi = false);
 
 #[bridge_io(TokioAsyncContext, ffi = false)]
 async fn RegistrationService_CreateSession(
@@ -124,7 +114,7 @@ async fn RegistrationService_SubmitCaptcha(
     service.0.lock().await.submit_captcha(&captcha_value).await
 }
 
-#[bridge_io(TokioAsyncContext, ffi = false, jni = false)]
+#[bridge_io(TokioAsyncContext, ffi = false)]
 async fn RegistrationService_RegisterAccount(
     service: &RegistrationService,
     register_account: &RegisterAccountRequest,
@@ -244,12 +234,12 @@ fn RegistrationSession_GetRequestedInformation(
     session.requested_information.iter().copied().collect()
 }
 
-#[bridge_fn(ffi = false, jni = false)]
+#[bridge_fn(ffi = false)]
 fn RegisterAccountRequest_Create() -> RegisterAccountRequest {
     RegisterAccountRequest(Some(RegisterAccountInner::default()).into())
 }
 
-#[bridge_fn(ffi = false, jni = false)]
+#[bridge_fn(ffi = false)]
 fn RegisterAccountRequest_SetSkipDeviceTransfer(register_account: &RegisterAccountRequest) {
     register_account
         .0
@@ -260,7 +250,7 @@ fn RegisterAccountRequest_SetSkipDeviceTransfer(register_account: &RegisterAccou
         .device_transfer = Some(libsignal_net::registration::SkipDeviceTransfer);
 }
 
-#[bridge_fn(ffi = false, jni = false)]
+#[bridge_fn(ffi = false)]
 fn RegisterAccountRequest_SetAccountPassword(
     register_account: &RegisterAccountRequest,
     account_password: String,
@@ -274,7 +264,22 @@ fn RegisterAccountRequest_SetAccountPassword(
         .account_password = account_password.into_boxed_str()
 }
 
-#[bridge_fn(ffi = false, jni = false)]
+// GCM is only used for Android.
+#[bridge_fn(ffi = false, node = false)]
+fn RegisterAccountRequest_SetGcmPushToken(
+    register_account: &RegisterAccountRequest,
+    gcm_push_token: String,
+) {
+    register_account
+        .0
+        .lock()
+        .expect("not poisoned")
+        .as_mut()
+        .expect("not taken")
+        .message_notification = NewMessageNotification::Gcm(gcm_push_token)
+}
+
+#[bridge_fn(ffi = false)]
 fn RegisterAccountRequest_SetIdentityPublicKey(
     register_account: &RegisterAccountRequest,
     identity_type: AsType<ServiceIdKind, u8>,
@@ -288,7 +293,7 @@ fn RegisterAccountRequest_SetIdentityPublicKey(
 /// cbindgen: ignore
 type SignedPublicPreKey = SignedPreKeyBody<Box<[u8]>>;
 
-#[bridge_fn(ffi = false, jni = false)]
+#[bridge_fn(ffi = false)]
 fn RegisterAccountRequest_SetIdentitySignedPreKey(
     register_account: &RegisterAccountRequest,
     identity_type: AsType<ServiceIdKind, u8>,
@@ -299,7 +304,7 @@ fn RegisterAccountRequest_SetIdentitySignedPreKey(
     *account.signed_pre_keys.get_mut(identity_type.into_inner()) = Some(signed_pre_key);
 }
 
-#[bridge_fn(ffi = false, jni = false)]
+#[bridge_fn(ffi = false)]
 fn RegisterAccountRequest_SetIdentityPqLastResortPreKey(
     register_account: &RegisterAccountRequest,
     identity_type: AsType<ServiceIdKind, u8>,
@@ -312,7 +317,7 @@ fn RegisterAccountRequest_SetIdentityPqLastResortPreKey(
         .get_mut(identity_type.into_inner()) = Some(pq_last_resort_pre_key);
 }
 
-#[bridge_fn(ffi = false, jni = false)]
+#[bridge_fn(ffi = false)]
 fn RegistrationAccountAttributes_Create(
     recovery_password: Box<[u8]>,
     aci_registration_id: u16,